[BUGFIX] fix switching to shadowmap render slicer causing rasterization renderer to break.
Current fix simply rearranges order of renders however this does not address the root cause which appears to be due to overwriting of the global_data object in shader.
Could be due to incorrect descriptor set layouts or something else causing the aforementioned memory to be overwritten.
Check the quad renderer and pipeline.